requirements.txt
discord.py~=2.3.2
requests~=2.31.0
pytube~=15.0.0
colorlog~=6.8.2